+++ fluid/PaddleCV/image_classification/utils/__init__.py
@@ -1,1 +1,2 @@
 from .learning_rate import cosine_decay, lr_warmup
+from .fp16_utils import create_master_params_grads, master_param_to_train_param
